refact(viewer): 合并StudentStoryViewer和StoryViewer到StoryViewer
